refactor(dto): separate DTOs into request and response folders
