+++ Backend/src/main/java/com/soydz/presentation/dto/request/UserRequestDTO.java
@@ -1,4 +1,4 @@
-package com.soydz.presentation.dto;
+package com.soydz.presentation.dto.request;
 
 import com.soydz.persistence.entity.RoleEntity;
 import com.soydz.persistence.entity.UserEntity;
