+++ Backend/src/main/java/com/soydz/presentation/dto/response/UserMovieResponseDetailsDTO.java
@@ -1,6 +1,7 @@
-package com.soydz.presentation.dto;
+package com.soydz.presentation.dto.response;
 
 import com.soydz.persistence.entity.UserMovieEntity;
+import com.soydz.presentation.dto.MovieDTO;
 
 import java.time.LocalDate;
