+++ Backend/src/main/java/com/soydz/service/impl/UserMovieServiceImpl.java
@@ -4,7 +4,12 @@ import com.soydz.persistence.entity.MovieEntity;
 import com.soydz.persistence.entity.UserEntity;
 import com.soydz.persistence.entity.UserMovieEntity;
 import com.soydz.persistence.repository.UserMovieRepository;
-import com.soydz.presentation.dto.*;
+import com.soydz.presentation.dto.MovieDTO;
+import com.soydz.presentation.dto.request.UserMovieRequestDTO;
+import com.soydz.presentation.dto.response.UserMovieResponseDTO;
+import com.soydz.presentation.dto.response.UserMovieResponseDetailsDTO;
+import com.soydz.presentation.dto.response.UserResponseDTO;
+import com.soydz.presentation.dto.response.UserResponseMinimalDTO;
 import com.soydz.service.interfaces.MovieService;
 import com.soydz.service.interfaces.UserMovieService;
 import com.soydz.service.interfaces.UserService;
