+++ Backend/src/main/java/com/soydz/service/impl/UserServiceImpl.java
@@ -3,8 +3,8 @@ package com.soydz.service.impl;
 import com.soydz.persistence.entity.RoleEntity;
 import com.soydz.persistence.entity.UserEntity;
 import com.soydz.persistence.repository.UserRepository;
-import com.soydz.presentation.dto.UserRequestDTO;
-import com.soydz.presentation.dto.UserResponseDTO;
+import com.soydz.presentation.dto.request.UserRequestDTO;
+import com.soydz.presentation.dto.response.UserResponseDTO;
 import com.soydz.service.interfaces.RoleService;
 import com.soydz.service.interfaces.UserService;
 import com.soydz.util.ValidationUtils;
