+++ Backend/src/main/java/com/soydz/service/interfaces/UserService.java
@@ -1,8 +1,8 @@
 package com.soydz.service.interfaces;
 
 import com.soydz.persistence.entity.UserEntity;
-import com.soydz.presentation.dto.UserRequestDTO;
-import com.soydz.presentation.dto.UserResponseDTO;
+import com.soydz.presentation.dto.request.UserRequestDTO;
+import com.soydz.presentation.dto.response.UserResponseDTO;
 
 import java.util.List;
 import java.util.Optional;
